better service reloads.
it doesn't matter when or which we connect the reload to because it's run at the end.

By default, notifications are :delayed, that is they are queued up as they are triggered, and then executed at the very end of the Chef run. If you want to run an action immediately, use :immediately:
http://wiki.opscode.com/display/chef/Resources#Resources-Notifications